Separate Logic and View implementation.
Move GameLogic in separated module with one repository.

Add module as dependency with help of jitpack.
