--- src/main/java/model/theComponents/Hero.java
@@ -1,8 +1,0 @@
-package model.theComponents;
-
-public class Hero extends GameFigure {
-
-    public Hero(String name, int hitPoints) {
-        super(name, hitPoints);
-    }
-}
