--- src/main/java/model/theWorld/map/DungeonMapBuilder.java
@@ -1,8 +1,0 @@
-package model.theWorld.map;
-
-public class DungeonMapBuilder {
-
-    public DungeonMap generateDungeonMap(int rowSize, int columnSize) {
-        return new DungeonMap(rowSize, columnSize);
-    }
-}
