--- src/main/java/model/theWorld/map/Map.java
@@ -1,16 +1,0 @@
-package model.theWorld.map;
-
-import model.theWorld.room.Room;
-
-public interface Map<R extends Room>  {
-
-    void addRoomToMap(R room);
-
-    R[][] getMap();
-
-    R getRoom(int row, int column);
-
-    boolean hasRoomAt(int row, int column);
-
-    boolean isEmptyAt(int row, int column);
-}
